Fixed makefile and dataset header magic size

+++ src/dataset_hdr.h
@@ -2,7 +2,7 @@
 #define AVC_DS_HDR
 
 typedef struct {
-	uint32_t magic;
+	uint64_t magic;
 	uint8_t is_raw;
 } dataset_header_t;
